+++ src/luanode_net.h
@@ -58,7 +58,7 @@ private:
 
 private:
 	lua_State* m_L;
-	const unsigned long m_socketId;
+	const unsigned int m_socketId;
 	bool m_close_pending;
 	bool m_write_shutdown_pending;
 	unsigned long m_pending_writes;
